Tests/LibWeb/Text/expected/HTML/fragment-quirks-mode-table-in-p.txt

BackCompat
<p><table><tbody><tr><td>cell</td></tr></tbody></table></p>
